refactor(legend): extract use-legend-layers hook
