+++ src/containers/map/legend/index.tsx
@@ -1,148 +1,32 @@
 'use client';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useState } from 'react';
 
-import { trackEvent } from '@/lib/analytics/ga';
 import cn from '@/lib/classnames';
-
-import { useSyncActiveLayers } from '@/store/layers';
 
 import { AnimatePresence, motion } from 'motion/react';
 import { FaArrowDown, FaArrowUp } from 'react-icons/fa6';
 import { IconBaseProps } from 'react-icons/lib';
 
-import { useSyncLocation } from 'hooks/use-sync-location';
-
-import { useLocation } from '@/containers/datasets/locations/hooks';
-import { LocationTypes } from '@/containers/datasets/locations/types';
-import { NATIONAL_DASHBOARD_LOCATIONS } from '@/containers/layers/constants';
-import { widgets } from '@/containers/widgets/constants';
-
 import SortableList from '@/components/map/legend/sortable/list';
-import type { Layer } from 'types/layers';
-import type { WidgetTypes } from 'types/widget';
 
 import LegendItem from './item';
+import { useLegendLayers } from './use-legend-layers';
 
 const FaArrowDownIcon = FaArrowDown as unknown as (p: IconBaseProps) => JSX.Element;
 const FaArrowUpIcon = FaArrowUp as unknown as (p: IconBaseProps) => JSX.Element;
 
-const NATIONAL_DASHBOARD_PREFIX = 'mangrove_national_dashboard_layer_';
-
 const Legend = ({ embedded = false }: { embedded?: boolean }) => {
-  const [activeLayers, setActiveLayers] = useSyncActiveLayers();
-
-  const { type, id: locationId } = useSyncLocation();
-  const locationType = (type ?? 'worldwide') as LocationTypes;
-
-  const { data: locationData } = useLocation(locationId, locationType);
-  const iso = locationData?.iso;
-
-  const nationalLayerIdPrefix = iso ? `${NATIONAL_DASHBOARD_PREFIX}${iso}` : null;
-
-  // When the user navigates to another country, drop any national-dashboard
-  // layers belonging to the previous ISO so the URL state stays in sync and
-  // we don't try to render Mapbox sources from a different country.
-  useEffect(() => {
-    if (!iso) return;
-    setActiveLayers((prev) => {
-      const current = prev ?? [];
-      const next = current.filter(
-        (layer) =>
-          !layer.id.startsWith(NATIONAL_DASHBOARD_PREFIX) ||
-          layer.id.startsWith(`${NATIONAL_DASHBOARD_PREFIX}${iso}_`) ||
-          layer.id === `${NATIONAL_DASHBOARD_PREFIX}${iso}`
-      );
-      return next.length === current.length ? prev : next;
-    });
-  }, [iso, setActiveLayers]);
-
-  const filterLayersByLocationType = useCallback(
-    (widgetsConfig: WidgetTypes[], currentLocationType: string): string[] => {
-      const filteredIds: string[] = [];
-
-      widgetsConfig.forEach((widget) => {
-        if (widget.locationType?.includes(currentLocationType)) {
-          if (widget.layersIds) filteredIds.push(...widget.layersIds);
-          if (widget.contextualLayersIds) filteredIds.push(...widget.contextualLayersIds);
-          if (widget.subLayersIds) filteredIds.push(...widget.subLayersIds);
-        }
-      });
-
-      return filteredIds;
-    },
-    []
-  );
-
-  const filteredLayersIds = useMemo(
-    () => filterLayersByLocationType(widgets, locationType),
-    [filterLayersByLocationType, locationType]
-  );
-
-  const filteredLayers = useMemo(() => {
-    const matched = (activeLayers ?? []).filter((layer) => {
-      const isStandardLayer = filteredLayersIds.includes(layer.id);
-
-      const isCurrentNationalLayer =
-        !!nationalLayerIdPrefix &&
-        (layer.id === nationalLayerIdPrefix || layer.id.startsWith(`${nationalLayerIdPrefix}_`));
-
-      return isStandardLayer || isCurrentNationalLayer;
-    }) as Layer[];
-
-    // Collapse all active national-dashboard layers under a single LegendItem;
-    // the WidgetLegend component lists each source row.
-    const firstNationalIdx = matched.findIndex((layer) =>
-      layer.id.startsWith(NATIONAL_DASHBOARD_PREFIX)
-    );
-    return matched.filter(
-      (layer, idx) => !layer.id.startsWith(NATIONAL_DASHBOARD_PREFIX) || idx === firstNationalIdx
-    );
-  }, [activeLayers, filteredLayersIds, nationalLayerIdPrefix]);
+  const { legendLayers, handleChangeOrder } = useLegendLayers();
 
   const [isOpen, setIsOpen] = useState(true);
 
   const openLegend = useCallback(() => {
-    if (activeLayers?.length) setIsOpen(true);
-  }, [activeLayers]);
+    if (legendLayers?.length) setIsOpen(true);
+  }, [legendLayers]);
 
   const closeLegend = useCallback(() => {
     setIsOpen(false);
   }, []);
-
-  const handleChangeOrder = useCallback(
-    (order: string[]) => {
-      const reorderedLayers = order
-        .map((layerId) => activeLayers?.find((layer) => layer.id === layerId))
-        .filter(Boolean) as Layer[];
-
-      const activePlanetLayers = activeLayers?.filter((layer) => layer.id.includes('planet')) ?? [];
-
-      trackEvent('Layers order', {
-        category: 'Layers',
-        action: 'Drag and drop',
-        label: `change layers order`,
-      });
-
-      setActiveLayers([...reorderedLayers, ...activePlanetLayers]);
-    },
-    [activeLayers, setActiveLayers]
-  );
-
-  const activeLayerNoPlanet = useMemo(
-    () =>
-      filteredLayers?.filter(
-        (layer) => !layer.id.includes('planet') && !layer.id.includes('custom-area')
-      ),
-    [filteredLayers]
-  );
-
-  const legendLayers = useMemo(() => {
-    const isNationalDashboardLocation = NATIONAL_DASHBOARD_LOCATIONS.includes(locationId);
-
-    if (isNationalDashboardLocation) return activeLayerNoPlanet;
-
-    return activeLayerNoPlanet?.filter((layer) => !layer.id.startsWith(NATIONAL_DASHBOARD_PREFIX));
-  }, [activeLayerNoPlanet, locationId]);
 
   return (
     <div>
